# api/apispec.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Wallet Ledger Service API",
    "description": "Backend API for a digital wallet and financial transaction system. Supports wallet management, fund transfers, ledger tracking, audit logging, and asynchronous event processing. Includes Admin endpoints for transaction monitoring and audit review.",
    "version": "1.1.0",
    "contact": {
      "name": "Marccel Janara"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/wallets/me": {
      "get": {
        "tags": ["Wallet"],
        "summary": "Get current user's wallet",
        "description": "Retrieves wallet information and balance for the authenticated user.",
        "operationId": "getMyWallet",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Wallet retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Wallet retrieved successfully",
                  "data": {
                    "wallet_id": "WLT-3KF91XA2B7",
                    "user_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "balance": "1500000.00",
                    "currency": "IDR",
                    "created_at": "2026-05-28T15:00:00Z",
                    "updated_at": "2026-05-28T16:30:00Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Wallet not found",
                  "error_code": 404,
                  "errors": null
                }
              }
            }
          }
        }
      }
    },
...
    "/ledger/entries": {
      "get": {
        "tags": ["Ledger"],
        "summary": "Get ledger entries for current user's wallet",
        "description": "Retrieves paginated ledger entries (DEBIT/CREDIT) for the authenticated user's wallet. Supports filtering by entry type.",
        "operationId": "getLedgerEntries",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "entry_type",
            "in": "query",
            "description": "Filter by entry type",
            "schema": {
              "type": "string",
              "enum": ["DEBIT", "CREDIT"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ledger entries retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Ledger entries retrieved successfully",
                  "data": {
                    "entries": [
                      {
                        "entry_id": "LED-AX82KD91NR",
                        "transaction_id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
                        "wallet_id": "WLT-3KF91XA2B7",
                        "entry_type": "DEBIT",
                        "amount": "250000.00",
                        "created_at": "2026-05-28T17:00:00Z"
                      }
                    ],
                    "pagination": {
                      "current_page": 1,
                      "per_page": 10,
                      "total_items": 1,
                      "total_pages": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/audit/logs": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get audit logs for current user",
        "description": "Retrieves paginated audit logs for the authenticated user's own activities (login, transfer, topup, etc.).",
        "operationId": "getAuditLogs",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Audit logs retrieved successfully",
                  "data": {
                    "logs": [
                      {
                        "log_id": "AUD-K82NRX91QP",
                        "user_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "action": "TRANSFER",
                        "ip_address": "192.168.1.100",
                        "endpoint": "POST /api/v1/transfers",
                        "created_at": "2026-05-28T17:00:00Z"
                      }
                    ],
                    "pagination": {
                      "current_page": 1,
                      "per_page": 10,
                      "total_items": 1,
                      "total_pages": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT access token stored in an httpOnly cookie. JWT payload includes user role."
      }
    },
...
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": ["success", "message", "data"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "description": "Response payload. Can be an object, array, string, number, boolean, or null."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "message", "error_code", "errors"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Operation failed"
          },
          "error_code": {
            "type": "integer",
            "example": 400
          },
          "errors": {
            "description": "Array of validation error strings, a single error string, or null",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Unauthorized: please login first",
              "error_code": 401,
              "errors": null
            }
          }
        }
      },
...
    }
  }
}